Split ServerOperations into server controllers
Change serverOperstions to Singleton class
